Fully functional automated assignment
The classes for automatic assignment are now functioning
mostly as intended (not ruling out unknown bugs).

The automated assignment will churn through every peak
and perform a frequency lookup with splatalogue. The
viable candidates are screened on the basis of upper state
energy (relative to the experiment temperature) and on
elemental composition (only matching the experiment).