Added harmonic progression fitting
Low level function implemented for trying to fit
frequencies to linear/prolate molecule model.

The algorithm takes a set of frequencies and the maximum
value allowed for J. The frequencies are sorted in ascending
order, and every possible quantum number ordering is generated
and iterated over.

The successful fit with the lowest RMS is returned.