Bug fix and harmonic progression implementation
Fixed issue where U-lines are not properly removed after
finalize_assignments is run. Now the peaks attribute of
an AssignmentSession is renewed when the assignments are
finalized, and then dumped to disk.

Bars for U-lines are also added to the plot_assignments.

Testing harmonic progression, but is incredibly slow; need
to optimize after I know it works.