Changes to frequency cutoff for assignments
Homogenized all of the frequency cut offs to use the Session attributes freq_abs and freq_prox. The former specifies whether or not the absolute value of freq_prox is used as the cutoff.

This means that the same cutoffs are used across the board for all assignment functions.

The default settings now are to use absolute values, and lines within 100 kHz.