Updated method for peak and noise detection
The find_peaks method for AssignmentSession is now updated to perform a quasi-intelligent noise floor detection when a threshold is not given.

The detect_noise_floor method will perform a cursory peak detection at 1% of the maximum signal, and finds the biggest gap between spectral features to find a region to calculate noise statistics.

The peaks_df will now also have the intensities shifted down by the baseline level