Further optimizations, and constraints
The QuantumNumber class now has a j attribute, which specifies whether or not it's a molecular rotation. If so, it constrains the change allowed to 0 or +1 for now.

The random seed generation has also been moved to the top _rng function: it will still be independently called by each thread, but no longer zealously so.