Moved to version 4
Major change now is moving to an entirely new system for handling peaks in a spectrum. Now everything is based on a LineList object, which handles everything from U-lines to assignments to catalogs.

Prior to this, the process_* functions relied on searching with Pandas dataframes. Having moved to a custom data structure now that is better suited for the task, there is a remarkable amount of speed up in processing line lists.

This commit includes working peak detection and line processing, but the remaining book keeping functions like finalize_assignments need to be updated.