Modified method for process_linelist_batch and LineList.from_lin
The former case now adds support for .lin files, and will use the file extensions to infer what parser to use when creating the LineList object.

The argument for LineList.from_lin has been modified; used to be linpath and now is filepath to be consistent with the .from_catalog method.