Added preliminary support for multiple assignments
During the automated `process_linelist`, if there are multiple
coincident frequencies, these are stored in a list attribute of
the `Transition` object corresponding to that U-line. When
`finalize_assignments` is called, each assigned line is checked
to see if there are multiple frequencies. If there are, a warning
is issued to tell the user there are ambigious assignments remaining.