Rolled to version 4.3.2
This version implements a change to how multiassignment is
done. `process_linelist` will only add to the `multiple`
attribute of `Transition` when the new transition does not
belong to the same molecule.

In `finalize_assignments`, there were previously duplicate
entries in the `multiple` attribute of `Transition`, which
should now only return unique entries.