Performance improvements on non-Cython lineshapes
Turns out some of the lineshapes in `fast` are broken for
whatever reason. Instead of fixing the Cython code, it's
probably more beneficial to just provide JIT'd code.

`pyspectools.lineshapes` is now optimized with Numba JIT,
with fairly reasonable performance on large arrays, and with
large numbers of multiple Gaussians. From this point we will
begin to depreciate the Cython routines in favor of JIT