Implemented working molecule formatting
This implements some classes for different classes of molecules for use
in SPCAT. Basically, it provides a human interface that maps string
identifiers of Hamiltonian parameters to the coding needed for
SPCAT/SPFIT, in a user friendly and extendible fashion.

The main scope for this interface is to capture the tediousness
associated with 90% of SPCAT usage; for the remaining 10% that are
complicated stuff involving things like interactions, you're better off
coding those parameters yourself, although each class includes a
`custom_coding` arg that you can pass for more definitions.