fix: hardcoded value to make nitrogen nuclei work
The `quanta` key in the `var_template` controls the nuclear spins as
well; for molecules with one nitrogen (spin 1.5) the value of quanta
becomes 3 (I guess for a triplet multiplicity?).

Temporarily `format_var` is patched to only assume nitrogen, which is
suboptimal. I'll need to rework the `nuclei` property of molecules to
contain a dictionary mapping of nuclei number and spin.