feat: database add and retrieve functionality
The `to_molecule` static method of `MoleculeDatabase` now provides a way
to retrieve entries, ready for SPCAT usage. The way that the molecule
data is stored with `tinyDB` is in JSON, and so everything is stored as
key/value pairs. Need to write documentation for that in the next
commit.